+++ app/assets/javascripts/views/story_view.js
@@ -11,6 +11,8 @@ import StoryLabels from 'components/story/StoryLabels';
 import StoryTasks from 'components/story/StoryTasks';
 import TaskForm from 'components/tasks/TaskForm';
 import StoryAttachment from 'components/story/StoryAttachment';
+import StoryStateButtons from 'components/story/StoryStateButtons'
+import StoryEstimateButtons from 'components/story/StoryEstimateButtons'
 
 var Clipboard = require('clipboard');
 
@@ -481,6 +483,27 @@ module.exports = FormView.extend({
     } else {
       this.$el.removeClass('editing');
       this.$el.html(this.template({story: this.model, view: this}));
+
+      const stateButtons = this.$('[data-story-state-buttons]').get(0)
+      if(stateButtons) {
+        ReactDOM.render(
+          <StoryStateButtons
+            events={this.model.events()}
+          />,
+          stateButtons
+        )
+      }
+
+      const estimateButtons = this.$('[data-story-estimate-buttons]').get(0)
+      if(estimateButtons) {
+        ReactDOM.render(
+          <StoryEstimateButtons
+            points={this.model.point_values()}
+          />,
+          estimateButtons
+        )
+      }
+
       if (isGuest) { this.$el.find('.state-actions').find('.transition').prop('disabled', true) }
     }